AlphaZero

Modules

AlphaZero.MCTS — Module

A generic, standalone implementation of asynchronous Monte Carlo Tree Search. It can be used on any game that implements the GameInterface interface and with any external oracle.

AlphaZero.GameInterface — Module

A generic interface for zero-sum, symmetric games.

AlphaZero.Network — Module

A generic, framework agnostic interface for neural networks.