AlphaZero

Modules

AlphaZero.MCTS — Module

A generic, standalone implementation of asynchronous Monte Carlo Tree Search. It can be used on any game that implements the GameInterface interface and with any external oracle.

AlphaZero.GameInterface — Module

A generic interface for zero-sum, symmetric games.

AlphaZero.Network — Module

A generic, framework agnostic interface for neural networks.

Training Parameters

AlphaZero.MctsParams — Type

Parameters of an MCTS player.

Parameter	Type	Default

num_workers	Int	1

use_gpu	Bool	false

num_iters_per_turn	Int	-

cpuct	Float64	1.

temperature	StepSchedule{Float64}	StepSchedule(1.)

dirichlet_noise_nα	Float64	10.

dirichlet_noise_ϵ	Float64	0.