MCTS

A generic, standalone implementation of asynchronous Monte Carlo Tree Search. It can be used on any game that implements the GameInterface interface and with any external oracle.

Oracles

AlphaZero.MCTS.Oracle — Type

MCTS.Oracle{Game}

Abstract base type for an oracle. Oracles must implement MCTS.evaluate and MCTS.evaluate_batch.

AlphaZero.MCTS.evaluate — Function

MCTS.evaluate(oracle::Oracle, board, available_actions)

Evaluate a single board position (assuming white is playing).

Return a pair (P, V) where:

P is a probability vector on available actions

V is a scalar estimating the value or win probability for white.

AlphaZero.MCTS.evaluate_batch — Function

MCTS.evaluate_batch(oracle::Oracle, batch)

Evaluate a batch of board positions.

Expect a vector of (board, available_actions) pairs and return a vector of (P, V) pairs.

A default implementation is provided that calls MCTS.evaluate sequentially on each position.

MCTS.evaluate_batch(::AbstractNetwork, batch)

Evaluate a batch of positions. See MCTS.evaluate_batch(::MCTS.Oracle).

AlphaZero.MCTS.RolloutOracle — Type

MCTS.RolloutOracle{Game} <: MCTS.Oracle{Game}

This oracle estimates the value of a position by simulating a random game from it (a rollout). Moreover, it puts a uniform prior on available actions. Therefore, it can be used to implement the "vanilla" MCTS algorithm.

Environment

AlphaZero.MCTS.Env — Type

MCTS.Env{Game}(oracle; nworkers=1, fill_batches=false, cpuct=1.) where Game

Create and initialize an MCTS environment.

Arguments

oracle: external oracle

nworkers: numbers of asynchronous workers (see below)

fill_batches: if true, a constant batch size is enforced for evaluation requests, by completing batches with dummy entries if necessary

cpuct: exploration constant (in the UCT formula)

Asynchronous MCTS

If nworkers == 1, MCTS is run in a synchronous fashion and the oracle is invoked through MCTS.evaluate.

If nworkers > 1, nworkers asynchronous workers are spawned, along with an additional task to serve board evaluation requests. Such requests are processed by batches of size nworkers using MCTS.evaluate_batch.

AlphaZero.MCTS.explore! — Function

MCTS.explore!(env, state, nsims)

Run nsims MCTS iterations from state.

In case there are multiple workers, nsims is rounded up to the nearest integer multiple of the number of workers.

AlphaZero.MCTS.policy — Function

MCTS.policy(env, state; τ=1.)

Return the recommended stochastic policy on state, with temperature parameter equal to τ.

A call to this function must always be preceded by a call to MCTS.explore!.

AlphaZero.MCTS.reset! — Function

MCTS.reset!(env)

Empty the MCTS tree.

Profiling utilities

AlphaZero.MCTS.inference_time_ratio — Function

MCTS.inference_time_ratio(env)

Return the ratio of time spent by MCTS.explore! on position evaluation (through functions MCTS.evaluate or MCTS.evaluate_batch) since the environment's creation.

AlphaZero.MCTS.memory_footprint_per_node — Function

MCTS.memory_footprint_per_node(env)

Return an estimate of the memory footprint of a single node of the MCTS tree (in bytes).

AlphaZero.MCTS.approximate_memory_footprint — Function

MCTS.approximate_memory_footprint(env)

Return an estimate of the memory footprint of the MCTS tree (in bytes).

AlphaZero.MCTS.average_exploration_depth — Function

MCTS.average_exploration_depth(env)

Return the average number of nodes that are traversed during an MCTS iteration, not counting the root.