Misc

AlphaZero.Session — Type

Session{Env}

A basic user interface for AlphaZero environments.

AlphaZero.Session — Method

Session(Game, Network, params, netparams;
  dir="session", autosave=true, validation=nothing)

Create a new session.

AlphaZero.Session — Method

Session(Game, Network, dir; autosave=true, validation=nothing)

Load an existing session from a directory.

AlphaZero.think — Function

think(::AbstractPlayer, state, turn_number::Int)

Return an (action, π) pair where action is the chosen action and π a probability distribution over available actions.

Note that a does not have to be drawn from π.

AlphaZero.pit — Function

pit(handler, baseline, contender, ngames [, reset_period])

Evaluate two players against each other on a series of games, alternating colors.

Arguments

handler: this function is called after each simulated game with two arguments: the game number i and the collected reward z for the contender player

baseline, contender :: AbstractPlayer

ngames: number of games to play

reset_period: if set, players are reset every reset_period games