Benchmark

AlphaZero.Benchmark — Module

Utilities to evaluate players against one another.

Duels

AlphaZero.Benchmark.Duel — Type

Benchmark.Duel(player, baseline; num_games)

Specify a duel that consists in num_games games between player and baseline, each of them of type Benchmark.Player.

Optional keyword arguments

reset_every: if set, the MCTS tree is reset every reset_mcts_every games to avoid running out of memory

color_policy has type ColorPolicy (ALTERNATE_COLORS) by default

AlphaZero.Benchmark.DuelOutcome — Type

Benchmark.DuelOutcome

The outcome of a duel between two players.

Fields

player and baseline are String fields containing the names of both players involved in the duel

avgz is the average reward collected by player

rewards is a vector containing all rewards collected by player (one per game played)

time is the amount of computing time took consumed by the duel, in seconds

AlphaZero.ColorPolicy — Type

@enum ColorPolicy ALTERNATE_COLORS BASELINE_WHITE CONTENDER_WHITE

Policy for attributing colors in a duel between a baseline and a contender.

Players

AlphaZero.Benchmark.Player — Type

Benchmark.Player

Abstract type for a player that can be featured in a benchmark duel.

AlphaZero.Benchmark.Full — Type

Benchmark.Full(params) <: Benchmark.Player

Full AlphaZero player that combines MCTS with the learnt network.

Constructor argument params has type MctsParams.

AlphaZero.Benchmark.NetworkOnly — Type

Benchmark.NetworkOnly(params) <: Benchmark.Player

Player that uses the policy output by the learnt network directly, instead of relying on MCTS.

Constructor argument params has type MctsParams, but only the following fields are used: use_gpu, temperature, dirichlet_noise_nα and dirichlet_noise_ϵ.

AlphaZero.Benchmark.MctsRollouts — Type

Benchmark.MctsRollouts(params) <: Benchmark.Player

Pure MCTS baseline that uses rollouts to evaluate new positions.

Constructor argument params has type MctsParams.

AlphaZero.Benchmark.MinMaxTS — Type

Benchmark.MinMaxTS(;depth, random_ϵ=0.) <: Benchmark.Player

Minmax baseline, which plans at depth depth and selects a random action with probability random_ϵ for exploration.