Game Interface

AlphaZero.GameInterface — Module

A generic interface for two-players, symmetric, zero-sum games.

Types

AlphaZero.GameInterface.AbstractGame — Type

AbstractGame

Abstract base type for a game state.

Constructors

Any subtype Game must implement Base.copy along with the following constructors:

Game()

Return the initial state of the game.

Game(board, white_playing=true)

Return the unique state specified by a board and a current player. By convention, the first player to play is called white and the other is called black.

AlphaZero.GameInterface.Board — Function

Board(Game::Type{<:AbstractGame})

Return the board type corresponding to Game.

Board objects must be persistent or appear as such.

AlphaZero.GameInterface.Action — Function

Action(Game::Type{<:AbstractGame})

Return the action type corresponding to Game.

Actions must be "symmetric" in the following sense:

available_actions(s) ==
  available_actions(Game(board_symmetric(s), !white_playing(s)))

Game Functions

AlphaZero.GameInterface.white_playing — Function

white_playing(state::AbstractGame) :: Bool

Return true if white is to play and false otherwise.

AlphaZero.GameInterface.white_reward — Function

white_reward(state::AbstractGame)

Return nothing if the game hasn't ended. Otherwise, return a reward for the white player as a number between -1 and 1.

AlphaZero.GameInterface.board — Function

board(state::AbstractGame)

Return the game board.

AlphaZero.GameInterface.board_symmetric — Function

board_symmetric(state::AbstractGame)

Return the symmetric of the game board (where the roles of black and white are swapped).

AlphaZero.GameInterface.actions — Function

actions(::Type{<:AbstractGame})

Return the vector of all game actions.

AlphaZero.GameInterface.actions_mask — Function

actions_mask(state::AbstractGame)

Return a boolean mask indicating what actions are available from state.

The following identities must hold:

game_terminated(state) || any(actions_mask(state))
length(actions_mask(state)) == length(actions(typeof(state)))

AlphaZero.GameInterface.available_actions — Function

available_actions(state::AbstractGame)

Return the vector of all available actions.

A default implementation is provided based on actions and actions_mask.

AlphaZero.GameInterface.play! — Function

play!(state::AbstractGame, action)

Update the game state by making the current player perform action.

AlphaZero.GameInterface.heuristic_value — Function

heuristic_value(state::AbstractGame)

Return a heuristic estimate of the state value for the current player.

The given state must be nonfinal and returned values must belong to the $(-∞, ∞)$ interval. Also, implementations of this function must be antisymmetric in the sense that:

heuristic_value(s) ==
  - heuristic_value(Game(board_symmetric(s), white_playing(s)))

This function is not needed by AlphaZero but it is useful for building baselines such as minmax players.

AlphaZero.GameInterface.vectorize_board — Function

vectorize_board(::Type{<:AbstractGame}, board) :: Vector{Float32}

Return a vectorized representation of a board.

AlphaZero.GameInterface.symmetries — Function

symmetries(::Type{G}, board) where {G <: AbstractGame}

Return the vector of all pairs (b, σ) where:

b is the image of board by a nonidentical symmetry

σ is the associated actions permutation, as an integer vector of size num_actions(Game).

Interface for Interactive Tools

AlphaZero.GameInterface.action_string — Function

action_string(::Type{<:AbstractGame}, action) :: String

Return a human-readable string representing the provided action.

AlphaZero.GameInterface.parse_action — Function

parse_action(::Type{<:AbstractGame}, str::String)

Return the action described by string str or nothing if str does not denote a valid action.

AlphaZero.GameInterface.read_state — Function

read_state(::Type{G}) where G <: AbstractGame :: Union{G, Nothing}

Read a state description from the standard input. Return the corresponding state or nothing in case of an invalid input.

AlphaZero.GameInterface.print_state — Function

print_state(state::AbstractGame)

Print a state on the standard output.

Utilities

AlphaZero.GameType — Function

GameType(T)

Return the Game type associated with an object.