Misc

AlphaZero.Session — Type

Session{Env}

A basic user interface for AlphaZero environments.

AlphaZero.Session — Method

Session(::Type{Game}, ::Type{Network}, params, netparams)

Create a new session.

AlphaZero.Session — Method

Session(::Type{Game}, ::Type{Network}, dir::String)

Load an existing session from a directory.

AlphaZero.think — Function

think(::AbstractPlayer, state, turn_number::Int)

Return an (a, π) pair where a is the chosen action and π a probability distribution over available actions.

Note that a does not have to be drawn from π.

AlphaZero.pit — Function

pit(handler, baseline, contender, ngames)

Evaluate two players against each other on a series of games.

Arguments

handler: this function is called after each simulated game with two arguments: the game number i and the collected reward z for the contender player

baseline, contender :: AbstractPlayer

ngames: number of games to play

Optional keyword arguments

reset_every: if set, players are reset every reset_every games

color_policy: determine the color attribution policy, which is ALTERNATE_COLORS by default