Miscellaneous

We gather on this page the documentation of internal utility functions that are omitted from the manual for brievety.

Batchifying oracles

AlphaZero.Batchifier — Module

Utilities for batchifying oracle evaluation.

The typical workflow should be the following:

Launch an inference server using launch_server to obtain a request channel.

Create several BatchedOracle that send queries along this channel.

When a client to the server is done, it calls client_done! so that the server does not have to wait for its next query anymore.

AlphaZero.Batchifier.launch_server — Function

launch_server(f; num_workers, batch_size)

Launch an inference requests server.

Arguments

f is the function to be evaluated (e.g. a neural network). It takes a batch of inputs and returns a batch of results of similar size.

num_workers is the number of workers that are expected to query the server.

batch_size corresponds to the batch size that is used to evaluate f. Note that one must have batch_size <= num_workers

How to use

This function returns a channel along which workers can send queries. A query can be either:

- `:none` when a worker is done and about to terminate
- a named tuple with fields `query` (the input to be given to `f`) and
  `answer_channel` (the channel the sever must use to return its answer).

The server stops automatically after all workers send :none.

AlphaZero.Batchifier.client_done! — Function

client_done!(reqc)

This function is to be called every time a client to the inference server identified by request channel reqc is done and won't send queries anymore.

It should be called $n$ times in total, where $n$ is the number of workers that was indicated to launch_server.

Note

This function does not take a BatchedOracle as an argument as there can be several oracles per worker (e.g. a worker can simulate games between two players that each rely on a neural network).

AlphaZero.Batchifier.BatchedOracle — Type

BatchedOracle(reqc, preprocess=(x->x))

Create an oracle that delegates its job to an inference server.

When called on a state, this oracle sends a query to the server identified by request channel reqc. This call is blocking until every other active worker also sends its query.

A preprocess function can be provided to ransform the passed state before it is sent to the server as a query.

KnetLib and FluxLib

AlphaZero.FluxLib — Module

This module provides utilities to build neural networks with Flux, along with a library of standard architectures.

AlphaZero.FluxLib.FluxNetwork — Type

FluxNetwork <: AbstractNetwork

Abstract type for neural networks implemented using the Flux framework.

The regularized_params_ function must be overrided for all layers containing parameters that are subject to regularization.

Provided that the above holds, FluxNetwork implements the full network interface with the following exceptions: Network.HyperParams, Network.hyperparams, Network.forward and Network.on_gpu.

AlphaZero.FluxLib.TwoHeadNetwork — Type

TwoHeadNetwork <: FluxNetwork

An abstract type for two-head neural networks implemented with Flux.

Subtypes are assumed to have fields hyper, gspec, common, vhead and phead. Based on those, an implementation is provided for Network.hyperparams, Network.game_spec, Network.forward and Network.on_gpu, leaving only Network.HyperParams to be implemented.

Utilities

AlphaZero.Util.apply_temperature — Method

apply_temperature(π, τ)

Apply temperature τ to probability distribution π. Handle the limit case where τ=0.

AlphaZero.Util.cycle_iterator — Method

cycle_iterator(iterator)

Generate an infinite cycle from an iterator.

AlphaZero.Util.fix_probvec — Method

fix_probvec(π)

Convert probability vector π to type Vector{Float32} and renormalize it if necessary.

This is useful as Distributions.isprobvec can be picky about its input when it does not sum to one due to numerical errors.

AlphaZero.Util.generate_update_constructor — Method

generate_update_constructor(T)

Generate a new constructor for immutable type T that enables copying an existing structure while only updating a subset of its fields.

For example, given the following struct:

struct Point
  x :: Int
  y :: Int
end

The generated code is equivalent to:

Point(pt; x=pt.x, y=pt.y) = Point(x, y)

** This function may be deprecated in favor of Setfield.jl in the future.**

AlphaZero.Util.mapreduce — Method

mapreduce(make_worker, args, num_workers, combine, init)

In spirit, this computes reduce(combine, map(f, args); init) on num_workers where f is defined below.

The make_worker function must create a worker w with two fields:

a process function such that w.process(x) evaluates to f(x)

a terminate function to be called with no argument when the worker terminates.

This function only spawns workers on background threads (with id greater or equal than 1).

Note

This function makes one call to combine per computed element and so it should only be used when the associated synchronization cost is small compared to the cost of computing individual elements.

AlphaZero.Util.momentum_smoothing — Method

Same smoothing function that is used by Tensorboard to smooth time series.

AlphaZero.Util.rand_categorical — Method

Draw a sample from a categorical distribution represented as a probability vector. See fix_probvec.

AlphaZero.Util.@printing_errors — Macro

@printing_errors expr

Evaluate expression expr while printing any uncaught exception on stderr. This is useful to avoid silent falure of concurrent tasks, as explained in this issue.