Training Reports

AlphaZero.Report — Module

Analytical reports generated during training, for debugging and hyperparameters tuning.

Learning Phase

AlphaZero.Report.Loss — Type

Report.Loss

Decomposition of the loss in a sum of terms (all have type Float32).

L is the total loss: L == Lp + Lv + Lreg + Linv

Lp is the policy cross-entropy loss term

Lv is the average value mean square error

Lreg is the L2 regularization loss term

Linv is the loss term penalizing the average weight put by the network on invalid actions

AlphaZero.Report.LearningStatus — Type

Report.LearningStatus

Statistics about the performance of the neural network on a subset of the memory buffer.

loss: detailed loss on the samples, as an object of type Report.Loss

Hp: average entropy of the $π$ component of samples (MCTS policy); this quantity is independent of the network and therefore constant during a learning iteration

Hpnet: average entropy of the network's prescribed policy on the samples

AlphaZero.Report.Checkpoint — Type

Report.Checkpoint

Report generated after a checkpoint evaluation.

epoch_id: number of epochs after which the checkpoint was computed

reward: average reward collected by the contender network

nn_replaced: true if the current best neural network was updated after the checkpoint

AlphaZero.Report.Epoch — Type

Report.Epoch

Report generated after each learning epoch

status_after: learning status after the epoch, as an object of type LearningStatus

AlphaZero.Report.Learning — Type

Report.Learning

Report generated at the end of the learning phase of an iteration.

time_convert, time_loss, time_train and time_eval are the amounts of time (in seconds) spent at converting the samples, computing losses, performing gradient updates and evaluating checkpoints respectively

initial_status: status before the learning phase, as an object of type Report.LearningStatus

epochs: vector of Report.Epoch reports

checkpoints: vector of Report.Checkpoint reports

nn_replaced: true if the best neural network was replaced

Memory Analysis Phase

AlphaZero.Report.Samples — Type

Report.Samples

Statistics about a set of samples, as collected during memory analysis.

num_samples: total number of samples

num_boards: number of distinct board positions

Wtot: total weight of the samples

status: Report.LearningSTatus statistics of the current network on the samples

AlphaZero.Report.StageSamples — Type

Report.StageSamples

Statistics for the samples corresponding to a particular game stage, as collected during memory analysis.

mean_remaining_length: average number of remaining moves until the end of the game; this quantity defines a game stage

samples_stats: samples statistics, as an object of type Report.Samples

AlphaZero.Report.Memory — Type

Report.Memory

Report generated by the memory analysis phase of an iteration. It features samples statistics for

the whole memory buffer (all_samples :: Report.Samples)

the samples collected during the last self-play iteration (latest_batch :: Report.Samples)

the subsets of the memory buffer corresponding to different game stages: (per_game_stage :: Vector{Report.StageSamples})

See MemAnalysisParams.

Self-Play Phase

AlphaZero.Report.SelfPlay — Type

Report.SelfPlay

Report generated after the self-play phase of an iteration.

inference_time_ratio: see MCTS.inference_time_ratio