Sessions

AlphaZero.Session — Type

Session{Env}

A wrapper on an AlphaZero environment that adds features such as:

Logging and plotting

Loading and saving of environments

In particular, it implements the Handlers interface.

Public fields

env::Env is the environment wrapped by the session

report is the current session report, with type SessionReport

Constructors

AlphaZero.Session — Method

Session(::Type{Game}, ::Type{Net}, params, netparams) where {Game, Net}

Create a new session given some parameters, or load it from disk if it already exists.

Arguments

Game is the type ot the game that is being learnt

Net is the type of the network that is being used

params has type Params

netparams has type Network.HyperParams(Net)

Optional keyword arguments

dir: session directory in which all files and reports are saved; this argument is either a string or nothing (default), in which case the session won't be saved automatically and no file will be generated

autosave: if set to false, the session won't be saved automatically nor any file will be generated (default is true)

nostdout: disables logging on the standard output when set to true (default is false)

benchmark: vector of Benchmark.Duel to be used as a benchmark (default is [])

load_saved_params: if set to true, load the training parameters from the session directory (if present) rather than using the params argument (default is false)

AlphaZero.Session — Method

Session(::Type{Game}, ::Type{Network}, dir::String) where {Game, Net}

Load an existing session from a directory.

This constructor features the optional keyword arguments autosave, nostdout and benchmark.

AlphaZero.Session — Method

Session(::Type{Game}, ::Type{Net}, params, netparams) where {Game, Net}

Create a new session given some parameters, or load it from disk if it already exists.

Arguments

Game is the type ot the game that is being learnt

Net is the type of the network that is being used

params has type Params

netparams has type Network.HyperParams(Net)

Optional keyword arguments

dir: session directory in which all files and reports are saved; this argument is either a string or nothing (default), in which case the session won't be saved automatically and no file will be generated

autosave: if set to false, the session won't be saved automatically nor any file will be generated (default is true)

nostdout: disables logging on the standard output when set to true (default is false)

benchmark: vector of Benchmark.Duel to be used as a benchmark (default is [])

load_saved_params: if set to true, load the training parameters from the session directory (if present) rather than using the params argument (default is false)

Session(::Type{Game}, ::Type{Network}, dir::String) where {Game, Net}

Load an existing session from a directory.

This constructor features the optional keyword arguments autosave, nostdout and benchmark.

Session(env::Env [, dir])

Create a session from an initial environment.

The iteration counter of the environment must be equal to 0

If a session directory is provided, this directory must not exist yet

This constructor features the optional keyword arguments autosave, nostdout and benchmark.

Session Reports

AlphaZero.SessionReport — Type

SessionReport

The full collection of benchmarks and statistics collected during a training session.

Fields

iterations: vector of $n$ iteration reports with type Report.Iteration

benchmark: vector of n+1 benchmark reports with type Benchmark.Report