Game Interface

AlphaZero.GameInterface — Module

A generic interface for two-players, symmetric, zero-sum games.

Interface Summary

Types, traits and constructors

Game()

Game(board)

Game(board, white_playing)

Base.copy(game)

Board(Game)

Action(Game)

Game functions

white_playing(state)

white_reward(state)

board(state)

board_symmetric(state)

available_actions(state)

play!(state, action)

Machine learning interface

vectorize_board(Game, board)

num_actions(Game)

action_id(Game, action)

action(Game, action_id)

Interface for interactive tools

action_string(Game, action)

parse_action(Game, str)

read_state(Game)

print_state(state)

Types

AlphaZero.GameInterface.AbstractGame — Type

AbstractGame

Abstract base type for a game state.

Constructors

Any subtype Game must implement the following constructors:

Game()

Return the initial state of the game.

Game(board, white_playing=true)

Return the unique state specified by a board and a current player. By convention, the first player to play is called white and the other is called black.

Base.copy — Method

Base.copy(::AbstractGame)

Return a fresh copy of a game state.

AlphaZero.GameInterface.Board — Function

Board(Game::Type{<:AbstractGame})

Return the board type corresponding to Game.

Board objects must be persistent or appear as such.

AlphaZero.GameInterface.Action — Function

Action(Game::Type{<:AbstractGame})

Return the action type corresponding to Game.

Actions must be "symmetric" in the following sense:

available_actions(s) ==
  available_actions(Game(board_symmetric(s), !white_playing(s)))

Game Functions

AlphaZero.GameInterface.white_playing — Function

white_playing(state::AbstractGame) :: Bool

Return true if white is to play and false otherwise.

AlphaZero.GameInterface.white_reward — Function

white_reward(state::AbstractGame)

Return nothing if the game hasn't ended. Otherwise, return a reward for the white player as a number between -1 and 1.

AlphaZero.GameInterface.board — Function

board(state::AbstractGame)

Return the game board.

AlphaZero.GameInterface.board_symmetric — Function

board_symmetric(state::AbstractGame)

Return the symmetric of the game board (where the roles of black and white are swapped).

AlphaZero.GameInterface.available_actions — Function

available_actions(state::AbstractGame)

Return the vector of all available actions, which must be nonempty if isnothing(white_reward(state)).

AlphaZero.GameInterface.play! — Function

play!(state::AbstractGame, action)

Update the game state by making the current player perform action.

AlphaZero.GameInterface.heuristic_value — Function

heuristic_value(state::AbstractGame)

Return a heuristic estimate of the state value for the current player.

The given state must be nonfinal and returned values must belong to the $(-∞, ∞)$ interval. Also, implementations of this function must be antisymmetric in the sense that:

heuristic_value(s) ==
  - heuristic_value(Game(board_symmetric(s), white_playing(s)))

This function is not needed by AlphaZero but it is useful for building baselines such as minmax players.

Machine Learning Interface

AlphaZero.GameInterface.vectorize_board — Function

vectorize_board(::Type{<:AbstractGame}, board) :: Vector{Float32}

Return a vectorized representation of a board.

AlphaZero.GameInterface.num_actions — Function

num_actions(::Type{<:AbstractGame}) :: Int

Return the total number of actions for a game.

AlphaZero.GameInterface.action_id — Function

action_id(G::Type{<:AbstractGame}, action) :: Int

Map each action to a unique number in the range 1:num_actions(G).

AlphaZero.GameInterface.action — Function

action(::Type{<:AbstractGame}, Int)

Inverse function of action_id.

Map an action identifier to an actual action.

Interface for Interactive Tools

AlphaZero.GameInterface.action_string — Function

action_string(::Type{<:AbstractGame}, action) :: String

Return a human-readable string representing the provided action.

AlphaZero.GameInterface.parse_action — Function

parse_action(::Type{<:AbstractGame}, str::String)

Return the action described by string str or nothing if str does not denote a valid action.

AlphaZero.GameInterface.read_state — Function

read_state(::Type{G}) where G <: AbstractGame :: Union{G, Nothing}

Read a state description from the standard input. Return the corresponding state or nothing in case of an invalid input.

AlphaZero.GameInterface.print_state — Function

print_state(state::AbstractGame)

Print a state on the standard output.