Network Interface

AlphaZero.Network — Module

A generic, framework agnostic interface for neural networks.

Mandatory interface

AlphaZero.Network.AbstractNetwork — Type

AbstractNetwork{Game} <: MCTS.Oracle{Game}

Abstract base type for a neural network.

Constructor

Any subtype Network must implement the following constructor:

Network(hyperparams)

where the expected type of hyperparams is given by HyperParams(Network).

AlphaZero.Network.HyperParams — Function

HyperParams(::Type{<:AbstractNetwork})

Return the hyperparameter type associated with a given network type.

AlphaZero.Network.hyperparams — Function

hyperparams(::AbstractNetwork)

Return the hyperparameters of a network.

AlphaZero.Network.forward — Function

forward(::AbstractNetwork, boards)

Compute the forward pass of the network on a batch of inputs.

Expect a Float32 tensor boards whose batch dimension is the last one.

Return a (P, V) triple where:

P is a matrix of size (num_actions, batch_size). It is allowed to put weight on invalid actions (see evaluate).

V is a row vector of size (1, batch_size)

AlphaZero.Network.train! — Function

train!(::AbstractNetwork, opt::OptimiserSpec, loss, data, learning_rate)

Train a given network on data.

AlphaZero.Network.set_test_mode! — Function

set_test_mode!(mode=true)

Put a network in test mode or in training mode. This is relevant for networks featuring layers such as batch normalization layers.

AlphaZero.Network.params — Function

params(::AbstractNetwork)

Return the collection of trainable parameters of the network.

AlphaZero.Network.regularized_params — Function

regularized_params(::AbstractNetwork)

Return the collection of regularized parameters of a network. This usually excludes neuron's biases.

Conversion and copy

Base.copy — Method

Base.copy(::AbstractNetwork)

Return a copy of the given network.

AlphaZero.Network.to_gpu — Function

to_gpu(::AbstractNetwork)

Return a copy of the given network that has been transferred to the GPU if one is available. Otherwise, return the given network untouched.

AlphaZero.Network.to_cpu — Function

to_cpu(::AbstractNetwork)

Return a copy of the given network that has been transferred to the CPU or return the given network untouched if it is already on CPU.

AlphaZero.Network.on_gpu — Function

on_gpu(::AbstractNetwork)

Test if a network is located on GPU.

AlphaZero.Network.convert_input — Function

convert_input(::AbstractNetwork, input)

Convert an array (or number) to the right format so that it can be used as an input by a given network.

AlphaZero.Network.convert_output — Function

convert_output(::AbstractNetwork, output)

Convert an array (or number) produced by a neural network to a standard CPU array (or number) type.

Misc

AlphaZero.Network.gc — Function

gc(::AbstractNetwork)

Perform full garbage collection.

Derived functions

Evaluation function

AlphaZero.Network.evaluate — Function

evaluate(network::AbstractNetwork, boards, action_masks)

Evaluate a batch of board positions. This function is a wrapper on forward that puts a zero weight on invalid actions.

Arguments

boards is a tensor whose last dimension has size bach_size

action_masks is a binary matrix of size (num_actions, batch_size)

Return

Return a (P, V, Pinv) triple where:

P is a matrix of size (num_actions, batch_size).

V is a row vector of size (1, batch_size).

Pinv is a row vector of size (1, batch_size) that indicates the total probability weight put by the network on invalid actions for each sample.

All tensors manipulated by this function have type Float32.

Oracle interface

AlphaZero.MCTS.evaluate — Method

MCTS.evaluate(oracle::Oracle, board, available_actions)

Evaluate a single board position (assuming white is playing).

Return a pair (P, V) where:

P is a probability vector on available actions

V is a scalar estimating the value or win probability for white.

MCTS.evaluate(::AbstractNetwork, board, available_actions)

Evaluate a single board position. See MCTS.evaluate(::MCTS.Oracle).

Warning

Evaluating a neural network on a single sample at a time is slow. When possible, use MCTS.evaluate_batch instead.

AlphaZero.MCTS.evaluate_batch — Method

MCTS.evaluate_batch(oracle::Oracle, batch)

Evaluate a batch of board positions.

Expect a vector of (board, available_actions) pairs and return a vector of (P, V) pairs.

A default implementation is provided that calls MCTS.evaluate sequentially on each position.

MCTS.evaluate_batch(::AbstractNetwork, batch)

Evaluate a batch of positions. See MCTS.evaluate_batch(::MCTS.Oracle).

Misc

AlphaZero.Network.num_parameters — Function

num_parameters(::AbstractNetwork)

Return the total number of parameters of a network.

AlphaZero.Network.num_regularized_parameters — Function

num_regularized_parameters(::AbstractNetwork)

Return the total number of regularized parameters of a network.

AlphaZero.Network.mean_weight — Function

mean_weight(::AbstractNetwork)

Return the mean absolute value of the regularized parameters of a network.

AlphaZero.Network.copy — Method

copy(::AbstractNetwork; on_gpu, test_mode)

A copy function that also handles CPU/GPU transfers and test/train mode switches.