Miscellaneous

We gather on this page the documentation of internal utility functions that are omitted from the manual for brievety.

AlphaZero.GameType — Function

GameType(T)

Return the Game type associated with an object (such as a network, a player, an MCTS environment...)

AlphaZero.Util.apply_temperature — Method

apply_temperature(π, τ)

Apply temperature τ to probability distribution π. Handle the limit case where τ=0.

AlphaZero.Util.batches — Method

batches(X, batchsize; partial=false)

Take a data tensor X and split it into batches of fixed size along the last dimension of X.

If partial=true and the number of samples in X is not a multiple of batchsize, then an additional smaller batch is added at the end (otherwise, it is discarded).

AlphaZero.Util.concat_columns — Method

concat_columns(cols) == hcat(cols...) # but faster

AlphaZero.Util.fix_probvec — Method

fix_probvec(π)

Convert probability vector π to type Vector{Float32} and renormalize it if necessary.

This is useful as Distributions.isprobvec can be picky about its input when it does not sum to one due to numerical errors.

AlphaZero.Util.generate_update_constructor — Method

generate_update_constructor(T)

Generate a new constructor for immutable type T that enables copying an existing structure while only updating a subset of its fields.

For example, given the following struct:

struct Point
  x :: Int
  y :: Int
end

Th generated code is equivalent to:

Point(pt; x=pt.x, y=pt.y) = Point(x, y)

AlphaZero.Util.momentum_smoothing — Method

Same smoothing function that is used by Temsorboard to smooth time series.

AlphaZero.Util.rand_categorical — Method

Draw a sample from a categorical distribution represented as a probability vector. See fix_probvec.

AlphaZero.Util.random_batches — Method

random_batches(convert, data::Tuple, batchsize; partial=false)

Take a tuple of data tensors, shuffle its samples according to a random permutation and split them into a sequence of minibatches.

The result is a lazy iterator that calls convert on the tensors of each new batch right before returning it. The convert function is typically used to transfer data onto the GPU.

Note

In the future, it may be good to deprecate this function along with random_batches_stream and use a standard solution instead, such as Flux.DataLoader.

AlphaZero.Util.random_batches_stream — Method

random_batches_stream(convert, data::Tuple, batchsize)

Generate an infinite stateful iterator of random batches by calling random_batches repeatedly. Every sample is guaranteed to be drawn exactly once per epoch.

AlphaZero.Util.superpose — Method

superpose(xs) == cat(xs..., dims=ndims(first(xs))+1) # but faster

AlphaZero.Util.@printing_errors — Macro

@printing_errors expr

Evaluate expression expr while printing any uncaught exception on stderr.

This is useful to avoid silent falure of concurrent tasks, as explained in this issue.